requirements.txt
flask
flask-smorest
python-dotenv
sqlachemy
flask-sqlalchemy
Flask-JWT-Extended
flask-migrate